+++ example/example_app/lib/anotherFile.dart
@@ -1,1 +1,0 @@
-enum HomeEvent { showInfo, showDiscover, showProfile }